Move PDF content extraction to a separate module
